+++ lib/Checked_oint.mli
@@ -69,6 +69,20 @@ type bitness =
 (** Represents an integer type. *)
 type int_ty = signedness * bitness [@@deriving eq, show, enumerate]
 
+include sig
+  val u8_int_ty : int_ty
+  val u16_int_ty : int_ty
+  val u32_int_ty : int_ty
+  val u64_int_ty : int_ty
+  val u128_int_ty : int_ty
+  val i8_int_ty : int_ty
+  val i16_int_ty : int_ty
+  val i32_int_ty : int_ty
+  val i64_int_ty : int_ty
+  val i128_int_ty : int_ty
+end
+[@@ocamlformat "module-item-spacing = compact"]
+
 (** Determines the type representation of a generic integer. *)
 val generic_int_ty : generic -> int_ty